Originally ADDLIB cmd is simulated by firstly extracting all members of archive into a temporary directory then inserting all or part of those into current archive.
The shortcoming of this approach is that if the archive in question
contains members with the same name, only one of those identical named
member remains after extraction. Also the code is lengthy and ugly.

This commit change ADDLIB to use the new added mode 'A', which does
not have above problems.